Refactor rcc.rs into multiple files
